# librefit-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: librefit-service API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
paths:
...
  /api/tracker/weight/create:
    post:
      tags:
      - Weight Tracker Resource
      operationId: createWeightTrackerEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeightTrackerEntry'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeightTrackerEntry'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tracker/weight/delete/{date}/{sequence}:
    delete:
      tags:
      - Weight Tracker Resource
      operationId: deleteWeightTrackerEntry
      parameters:
      - name: date
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      - name: sequence
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tracker/weight/last:
    get:
      tags:
      - Weight Tracker Resource
      operationId: findLastWeightTrackerEntry
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeightTrackerEntry'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tracker/weight/list/{dateFrom}/{dateTo}:
    get:
      tags:
      - Weight Tracker Resource
      operationId: listWeightTrackerEntriesRange
      parameters:
      - name: dateFrom
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      - name: dateTo
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeightTrackerEntry'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tracker/weight/list/{date}:
    get:
      tags:
      - Weight Tracker Resource
      operationId: listWeightTrackerEntries
      parameters:
      - name: date
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeightTrackerEntry'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tracker/weight/read/{date}/{sequence}:
    get:
      tags:
      - Weight Tracker Resource
      operationId: readWeightTrackerEntry
      parameters:
      - name: date
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      - name: sequence
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeightTrackerEntry'
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tracker/weight/update:
    put:
      tags:
      - Weight Tracker Resource
      operationId: updateWeightTrackerEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeightTrackerEntry'
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
...
components:
  schemas:
...
    ErrorDescription:
      required:
      - field
      - message
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      required:
      - errors
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDescription'
...
    LocalDate:
      format: date
      type: string
      example: 2022-03-10
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
    WeightTrackerEntry:
      required:
      - added
      type: object
      properties:
        userId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        added:
          $ref: '#/components/schemas/LocalDate'
        sequence:
          format: int64
          type: integer
          nullable: true
        amount:
          format: float
          minimum: 0
          type: number
        updated:
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          nullable: true
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT